Migrate the shared components to align with Phlex conversions
